requirements.txt
requests
beautifulsoup4
google-cloud-bigquery
pandas
python-dotenv
pyarrow
google-cloud